// frontend/src/api/generated/models/measurementResponseStateEnum.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Dekopen API
 * Authenticated tenant and engine API boundary.
 * OpenAPI spec version: 1.0.0
 */

/**
 * * `CLIENT_DECLARED` - CLIENT_DECLARED
 * * `SITE_RECTIFIED` - SITE_RECTIFIED
 * * `CONFIRMED` - CONFIRMED
 */
export type MeasurementResponseStateEnum =
  (typeof MeasurementResponseStateEnum)[keyof typeof MeasurementResponseStateEnum];

export const MeasurementResponseStateEnum = {
  CLIENT_DECLARED: "CLIENT_DECLARED",
  SITE_RECTIFIED: "SITE_RECTIFIED",
  CONFIRMED: "CONFIRMED",
} as const;