// frontend/src/api/generated/models/wallTypeEnum.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Dekopen API
 * Authenticated tenant and engine API boundary.
 * OpenAPI spec version: 1.0.0
 */

/**
 * * `MASONRY` - MASONRY
 * * `CONCRETE` - CONCRETE
 * * `PARTITION` - PARTITION
 * * `WOOD` - WOOD
 */
export type WallTypeEnum = (typeof WallTypeEnum)[keyof typeof WallTypeEnum];

export const WallTypeEnum = {
  MASONRY: "MASONRY",
  CONCRETE: "CONCRETE",
  PARTITION: "PARTITION",
  WOOD: "WOOD",
} as const;